extracted all config from the app level into a context, awesome cleanup
